Initial work on audio support via PulseAudio
Currently this only supports enumerating audio devices, and it does
not have much in terms of documentation or unit testing.

Changes to design:

- IAudioProvider has two new methods, StartAsync and StopAsync
  - This is to support APIs which have asynchronous initialization,
    such as PulseAudio.
  - APIs return ValueTask, as not all backends will require asynchronous
    initialization, so they can return synchronously without causing any
    overheads.
- IAudioProvider.RequestAudioDevice is now async, returning a
  ValueTask<IAudioDevice>
  - Again, this is to support APIs which require asynchronous
    initialization of devices, such as PulseAudio.
  - Likewise, not all backends will require asynchronous initialization
    so it returns ValueTask to allow those which are not asynchronous to
    use it.
- Added AudioDeviceType enum to discern recording from playback devices
  in IAudioDeviceOptions
  - This could be expanded in the future to further identify device
    types, e.g. recording devices which record system-wide sound output,
    or 'Stereo Mix' devices on Windows.

Open Questions:

Do we use IAsyncEnumerable for EnumerateAudioDevices?
In the PulseAudio implementation I have had to use a BlockingCollection
to implement it as PulseAudio returns the device list asynchronously.
Maybe we could change the interface to return a
IAsyncEnumerable<IAudioDeviceOptions> instead and then return some other
type, possibly one generated by the compiler.

Do we change EnumerateAudioDevices to return a different type from
RequestAudioDeviceAsync?
Currently, we use the same type to implement both. However, this leaves
us in a weird state where we're returning mutable information from a
method, and also accepting properties which may not even make sense (e.g.
Name) when requesting audio devices.